Move when-context handling to its own file
